only exportGraphToDot left for change and unuit-test

+++ src/navigator/matrix/Matrix.hpp
@@ -12,7 +12,7 @@ public:
 
   std::size_t getSideSize() const { return _sideSize; }
 
-  double& at(int row, int col) { return _matrix[row][col]; }
+  double& at(size_t row, size_t col) { return _matrix[row][col]; }
 };
